+++ src/move-ts-indexer/util/shared-interfaces.ts
@@ -3,3 +3,18 @@ export interface Reference {
   specifiers: string[];
   isExport?: boolean;
 }
+
+export interface FoundItem {
+  itemType:
+    | 'importPath'
+    | 'exportPath'
+    | 'class'
+    | 'selector'
+    | 'templateUrl'
+    | 'styleUrl'
+    | 'styleUrls'
+    | 'attributeInput';
+  itemText: string;
+  specifiers?: string[];
+  location: { start: number; end: number };
+}