+++ src/test/suite/index4mac.ts
@@ -1,5 +1,5 @@
 import { runMochaScenarioFile } from './helpers/run-mocha-scenario-file.function';
 
 export function run(): Promise<void> {
-  return runMochaScenarioFile('./suite/scenario5.test.js');
+  return runMochaScenarioFile('./suite/scenario4-mac.test.js');
 }