+++ src/test/suite/index4windows.ts
@@ -1,5 +1,5 @@
 import { runMochaScenarioFile } from './helpers/run-mocha-scenario-file.function';
 
 export function run(): Promise<void> {
-  return runMochaScenarioFile('./suite/scenario2.test.js');
+  return runMochaScenarioFile('./suite/scenario4-windows.test.js');
 }