+++ src/test/suite/index5mac.ts
@@ -1,5 +1,5 @@
 import { runMochaScenarioFile } from './helpers/run-mocha-scenario-file.function';
 
 export function run(): Promise<void> {
-  return runMochaScenarioFile('./suite/scenario1.test.js');
+  return runMochaScenarioFile('./suite/scenario5-mac.test.js');
 }