+++ src/test/suite/index6windows.ts
@@ -1,5 +1,5 @@
 import { runMochaScenarioFile } from './helpers/run-mocha-scenario-file.function';
 
 export function run(): Promise<void> {
-  return runMochaScenarioFile('./suite/scenario3.test.js');
+  return runMochaScenarioFile('./suite/scenario6-windows.test.js');
 }